requirements.txt
django-tailwind==4.0.1
